Adicionando paginação para a listagem de usuários

+++ rest-java/rest-java-ws/src/main/java/br/com/rbh/service/UsuarioService.java
@@ -9,21 +9,26 @@ import javax.persistence.TypedQuery;
 
 import br.com.rbh.dominio.Usuario;
 
-@Stateless
+//@Stateless
 public class UsuarioService {
 
-	@PersistenceContext
+	//@PersistenceContext
 	private EntityManager em;
 	
 	public Usuario get(Long id) {
 		return em.find(Usuario.class, id);
 	}
 	
-	public List<Usuario> consultar() {
+	public List<Usuario> consultar(Integer limit, Integer offset) {
 		StringBuilder hql = new StringBuilder();
 		hql.append("FROM Usuario");
 		
 		TypedQuery<Usuario> query = em.createQuery(hql.toString(), Usuario.class);
+		
+		if (limit != null && offset != null) {
+			query.setMaxResults(limit);
+			query.setFirstResult(offset);
+		}
 		
 		return query.getResultList();
 	}
@@ -38,7 +43,11 @@ public class UsuarioService {
 		return em.merge(usuario);
 	}
 	
-	public void remover(Usuario usuario) {
-		em.remove(usuario);
+	public void remover(Long id) {
+		Usuario usuario = get(id);
+
+		if (usuario != null) {
+			em.remove(usuario);
+		}
 	}
 }
